requirements.txt
flask
flask-SQLAlchemy
flask-migrate
flask-Cors
psycopg2-binary
python-dotenv
